feat/polishing courses crud

+++ src/main/java/com/texas/developers/ams/repo/CourseRepository.java
@@ -7,6 +7,6 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface CourseRepository extends JpaRepository<Course, Long> {
+public interface CourseRepository extends JpaRepository<Course, Integer> {
     Optional<Course> findCourseByCourseName(String courseName);
 }
